requirements.txt
# Ultra-minimal requirements for Render.com deployment
# Only packages with pre-built wheels - NO compilation needed

# Core API (Fast install, pre-built wheels)
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Utilities (pre-built)
requests==2.31.0
python-dateutil==2.8.2

# Simple data structures (no numpy)
# App will work with pure Python + pattern-based predictions
